Fix test_network
